Move search-metadata-previews to separate package
