+++ packages/search-metadata-previews/SnippetEditor/components/ReplacementVariableEditorStandalone.js
