+++ packages/search-metadata-previews/SnippetEditor/components/SettingsSnippetEditor.js
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 
 // Internal dependencies.
 import SettingsSnippetEditorFields from "./SettingsSnippetEditorFields";
-import ErrorBoundary from "../../../basic/ErrorBoundary";
+import ErrorBoundary from "yoast-components/composites/basic/ErrorBoundary";
 import {
 	replacementVariablesShape,
 	recommendedReplacementVariablesShape,
