+++ packages/search-metadata-previews/SnippetEditor/components/__mocks__/ReplacementVariableEditorStandalone.js
