+++ packages/search-metadata-previews/SnippetEditor/replaceText.js
