+++ packages/search-metadata-previews/SnippetEditor/selection.js
