+++ packages/search-metadata-previews/SnippetEditor/serialization.js
