+++ packages/search-metadata-previews/SnippetEditor/tests/ReplacementVariableEditorTest.js
