+++ packages/search-metadata-previews/SnippetPreview/components/FixedWidthContainer.js
