+++ packages/search-metadata-previews/SnippetPreview/components/ProgressBar.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import styled from "styled-components";
 
 /* Internal dependencies */
-import colors from "../../../../style-guide/colors.json";
+import { colors } from "yoast-components";
 
 /**
  * The ProgressBar component.
