+++ packages/search-metadata-previews/SnippetPreview/components/SnippetPreview.js
@@ -12,11 +12,10 @@ import { __ } from "@wordpress/i18n";
 
 // Internal dependencies.
 import FixedWidthContainer from "./FixedWidthContainer";
-import colors from "../../../../style-guide/colors";
-import ScreenReaderText from "../../../../a11y/ScreenReaderText";
+import { colors, getRtlStyle } from "yoast-components";
+import ScreenReaderText from "yoast-components/a11y/ScreenReaderText";
 import { DEFAULT_MODE, MODE_DESKTOP, MODE_MOBILE, MODES } from "../constants";
 import { angleLeft, angleRight } from "../../SnippetEditor/components/Shared";
-import { getRtlStyle } from "../../../../utils/helpers/styled-components";
 
 /*
  * These colors should not be abstracted. They are chosen because Google renders
