+++ packages/search-metadata-previews/index.js
@@ -1,0 +1,2 @@
+export * from "./SnippetEditor";
+export * from "./SnippetPreview";
