+++ packages/search-metadata-previews/jest/setupTests.js
@@ -1,0 +1,1 @@
+import "jest-styled-components";
