+++ packages/search-metadata-previews/package.json
@@ -4,5 +4,46 @@
   "main": "index.js",
   "repository": "https://github.com/yoast/javascript",
   "author": "Team Yoast",
-  "license": "GPL-3.0"
+  "license": "GPL-3.0",
+  "scripts": {
+    "test": "jest",
+    "lint": "grunt check"
+  },
+  "dependencies": {
+    "@wordpress/a11y": "^1.0.7",
+    "@wordpress/i18n": "^1.1.0",
+    "draft-js": "^0.10.5",
+    "draft-js-mention-plugin": "^3.0.4",
+    "draft-js-plugins-editor": "^2.0.4",
+    "draft-js-single-line-plugin": "^2.0.1",
+    "prop-types": "^15.6.0",
+    "styled-components": "^2.1.2",
+    "yoast-components": "^4.22.0",
+    "yoastseo": "^1.49.0",
+    "lodash": "^4.17.11"
+  },
+  "devDependencies": {
+    "react": "16.6.3",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
+    "enzyme-to-json": "^3.3.3",
+    "jest-styled-components": "^5.0.1",
+    "jest-mock": "^22.4.3",
+    "jest-runtime": "^22.4.3",
+    "babel-jest": "^22.4.3"
+  },
+  "jest": {
+    "testURL": "http://localhost",
+    "testRegex": ".*Test.js$",
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ],
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest"
+    },
+    "transformIgnorePatterns": [
+      "/node_modules/(?!yoastseo|lodash-es).+\\.js$"
+    ],
+    "setupTestFrameworkScriptFile": "<rootDir>/jest/setupTests.js"
+  }
 }
